adjusted input width on landing page
