+++ src/index.css
@@ -41,3 +41,4 @@
 .fill_arrow {
     fill: #048a81;
 }
+
